pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinicius Borges",email = "vinicius.borges@ericsson.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=24.3,<25.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
pytest-django = "^4.9.0"
mypy = "^1.14.1"
ipykernel = "^6.29.5"
factory-boy = "^3.3.3"
django-extensions = "^3.2.3"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py"]
addopts = [
    "-vv", 
    "-s",
    "--reuse-db",
    "-p no:warnings",
]


[tool.ruff]
exclude = [
    ".venv",
    "node_modules",
    "*.ipynb",
]

line-length = 88

target-version = "py312"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "I",
    "DJ",
]
ignore = ["DJ001"]

fixable = ["ALL"]
unfixable = []


dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"


[tool.ruff.lint.isort.sections]
"django" = ["django"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "first-party",
    "local-folder",
    "third-party",
]